the .card class on the Registration.module.css style sheet is overriding my styling on my card components so my formatting was overriden. If a regular style sheet is used on the Registration sheet or the card class is renamed it works
